Change issue number from integer to bigint

+++ db/migrate/20201022004221_create_issues.rb
@@ -1,7 +1,7 @@
 class CreateIssues < ActiveRecord::Migration[6.0]
   def change
     create_table :issues do |t|
-      t.integer :number, null: false
+      t.bigint :number, null: false
 
       t.string :title, null: false
       t.string :body
